backend/requirements.txt
aioredis==1.3.1
amqp==2.6.0
anyjson==0.3.3
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
Babel==2.8.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
cachetools==4.1.0
celery==4.4.5
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.6.0
devourer==0.5.0
diff-match-patch==20181111
Django==3.0.6
django-adminlte-ui==1.5.1
django-celery-beat==2.0.0
django-codemirror2==0.2
django-concurrency==2.1.1
django-cors-headers==3.2.1
django-countries==6.1.2
django-cte==1.1.4
django-import-export==2.2.0
django-js-asset==1.2.2
django-material-admin==1.7.3
django-money==1.1
django-mptt==0.9.1
django-phonenumber-field==4.0.0
django-rest-multiple-models==2.1.3
django-rest-passwordreset==1.1.0
django-river==3.2.0
django-simple-history==2.10.0
django-tenant-schemas==1.10.0
django-timezone-field==4.0
django-treebeard==4.3
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
djangorestframework-simplejwt==4.4.0
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.25.2
future==0.18.2
futures==3.1.1
gevent==20.5.0
greenlet==0.4.15
hiredis==1.0.1
httplib2==0.18.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
jdcal==1.4.1
kombu==4.6.10
MarkupPy==1.14
mock==4.0.2
monotonic==1.5
msgpack==0.6.2
oauth2client==1.5.2
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
ordered-set==4.0.2
phonenumbers==8.12.10
Pillow==7.1.2
protobuf==3.12.2
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyPEG2==2.15.2
python-crontab==2.5.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.2
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.1.1
service-identity==18.1.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
structlog==20.1.0
tablib==2.0.0
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.1.0
uWSGI>=2.0.18,<2.1
psycopg2-binary==2.8.6
